requirements.txt
streamlit==1.39.0
ultralytics==8.3.0
opencv-python-headless==4.10.0.84
numpy==2.0.2
pandas==2.2.2
pillow==10.4.0